Rename package genome
