+++ genetics/file.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package genome
+package genetics
 
 import (
 	"bufio"
